Store firing schedule state on rotating files
the standard filename still used, but each time it gets updated, a copy
is saved to a another file. If something bad happens while writting to
disk, the same data is stored in two files, making it unlikely for both
to be corrupted. If that happens, the previous version should still be
intact.